CAG_Project_API/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pypdf
google-generativeai
python-dotenv
